--- src/com/Super135/Lesson1/Jumpable.java
@@ -1,5 +1,0 @@
-package com.Super135.Lesson1;
-
-public interface Jumpable {
-    void jumpAction (double height);
-}
