--- src/com/Super135/Lesson1/Runnable.java
@@ -1,5 +1,0 @@
-package com.Super135.Lesson1;
-
-public interface Runnable {
-    void runAction(int length);
-}
